+++ frontend/src/metabase/new_model/components/NewModelOption.jsx
@@ -4,7 +4,7 @@ import cx from "classnames";
 import { Link } from "react-router";
 import { color } from "metabase/lib/colors";
 
-export default class NewDatasetOption extends Component {
+export default class NewModelOption extends Component {
   state = {
     hover: false,
   };
